Update BitBucket + GitLab Payloads

+++ gitlab/payload.go
@@ -54,15 +54,15 @@ type PushEventPayload struct {
 	After             string     `json:"after"`
 	Ref               string     `json:"ref"`
 	CheckoutSHA       string     `json:"checkout_sha"`
-	UserID            int        `json:"user_id"`
+	UserID            int64      `json:"user_id"`
 	UserName          string     `json:"user_name"`
 	UserEmail         string     `json:"user_email"`
 	UserAvatar        string     `json:"user_avatar"`
-	ProjectID         int        `json:"project_id"`
+	ProjectID         int64      `json:"project_id"`
 	Project           Project    `json:"Project"`
 	Repository        Repository `json:"repository"`
 	Commits           []Commit   `json:"commits"`
-	TotalCommitsCount int        `json:"total_commits_count"`
+	TotalCommitsCount int64      `json:"total_commits_count"`
 }
 
 // TagEventPayload contains the information for GitLab's tag push event
@@ -72,14 +72,14 @@ type TagEventPayload struct {
 	After             string     `json:"after"`
 	Ref               string     `json:"ref"`
 	CheckoutSHA       string     `json:"checkout_sha"`
-	UserID            int        `json:"user_id"`
+	UserID            int64      `json:"user_id"`
 	UserName          string     `json:"user_name"`
 	UserAvatar        string     `json:"user_avatar"`
-	ProjectID         int        `json:"project_id"`
+	ProjectID         int64      `json:"project_id"`
 	Project           Project    `json:"Project"`
 	Repository        Repository `json:"repository"`
 	Commits           []Commit   `json:"commits"`
-	TotalCommitsCount int        `json:"total_commits_count"`
+	TotalCommitsCount int64      `json:"total_commits_count"`
 }
 
 // WikiPageEventPayload contains the information for GitLab's wiki created/updated event
@@ -105,7 +105,7 @@ type PipelineEventPayload struct {
 type CommentEventPayload struct {
 	ObjectKind       string           `json:"object_kind"`
 	User             User             `json:"user"`
-	ProjectID        int              `json:"project_id"`
+	ProjectID        int64            `json:"project_id"`
 	Project          Project          `json:"project"`
 	Repository       Repository       `json:"repository"`
 	ObjectAttributes ObjectAttributes `json:"object_attributes"`
@@ -122,15 +122,15 @@ type BuildEventPayload struct {
 	Tag               bool        `json:"tag"`
 	BeforeSHA         string      `json:"before_sha"`
 	SHA               string      `json:"sha"`
-	BuildID           int         `json:"build_id"`
+	BuildID           int64       `json:"build_id"`
 	BuildName         string      `json:"build_name"`
 	BuildStage        string      `json:"build_stage"`
 	BuildStatus       string      `json:"build_status"`
 	BuildStartedAt    customTime  `json:"build_started_at"`
 	BuildFinishedAt   customTime  `json:"build_finished_at"`
-	BuildDuration     int         `json:"build_duration"`
+	BuildDuration     int64       `json:"build_duration"`
 	BuildAllowFailure bool        `json:"build_allow_failure"`
-	ProjectID         int         `json:"project_id"`
+	ProjectID         int64       `json:"project_id"`
 	ProjectName       string      `json:"project_name"`
 	User              User        `json:"user"`
 	Commit            BuildCommit `json:"commit"`
@@ -139,24 +139,24 @@ type BuildEventPayload struct {
 
 // Issue contais all of the GitLab issue information
 type Issue struct {
-	ID          int        `json:"id"`
+	ID          int64      `json:"id"`
 	Title       string     `json:"title"`
-	AssigneeID  int        `json:"assignee_id"`
-	AuthorID    int        `json:"author_id"`
-	ProjectID   int        `json:"project_id"`
+	AssigneeID  int64      `json:"assignee_id"`
+	AuthorID    int64      `json:"author_id"`
+	ProjectID   int64      `json:"project_id"`
 	CreatedAt   customTime `json:"created_at"`
 	UpdatedAt   customTime `json:"updated_at"`
-	Position    int        `json:"position"`
+	Position    int64      `json:"position"`
 	BranchName  string     `json:"branch_name"`
 	Description string     `json:"description"`
-	MilestoneID int        `json:"milestone_id"`
+	MilestoneID int64      `json:"milestone_id"`
 	State       string     `json:"state"`
-	IID         int        `json:"iid"`
+	IID         int64      `json:"iid"`
 }
 
 // Build contais all of the GitLab build information
 type Build struct {
-	ID            int           `json:"id"`
+	ID            int64         `json:"id"`
 	Stage         string        `json:"stage"`
 	Name          string        `json:"name"`
 	Status        string        `json:"status"`
@@ -199,30 +199,30 @@ type Commit struct {
 
 // BuildCommit contais all of the GitLab build commit information
 type BuildCommit struct {
-	ID          int        `json:"id"`
+	ID          int64      `json:"id"`
 	SHA         string     `json:"sha"`
 	Message     string     `json:"message"`
 	AuthorName  string     `json:"auuthor_name"`
 	AuthorEmail string     `json:"author_email"`
 	Status      string     `json:"status"`
-	Duration    int        `json:"duration"`
+	Duration    int64      `json:"duration"`
 	StartedAt   customTime `json:"started_at"`
 	FinishedAt  customTime `json:"finished_at"`
 }
 
 // Snippet contais all of the GitLab snippet information
 type Snippet struct {
-	ID              int        `json:"id"`
+	ID              int64      `json:"id"`
 	Title           string     `json:"title"`
 	Content         string     `json:"content"`
-	AuthorID        int        `json:"author_id"`
-	ProjectID       int        `json:"project_id"`
+	AuthorID        int64      `json:"author_id"`
+	ProjectID       int64      `json:"project_id"`
 	CreatedAt       customTime `json:"created_at"`
 	UpdatedAt       customTime `json:"updated_at"`
 	FileName        string     `json:"file_name"`
 	ExpiresAt       customTime `json:"expires_at"`
 	Type            string     `json:"type"`
-	VisibilityLevel int        `json:"visibility_level"`
+	VisibilityLevel int64      `json:"visibility_level"`
 }
 
 // User contais all of the GitLab user information
@@ -241,7 +241,7 @@ type Project struct {
 	GitSSSHURL        string `json:"git_ssh_url"`
 	GitHTTPURL        string `json:"git_http_url"`
 	Namespace         string `json:"namespace"`
-	VisibilityLevel   int    `json:"visibility_level"`
+	VisibilityLevel   int64  `json:"visibility_level"`
 	PathWithNamespace string `json:"path_with_namespace"`
 	DefaultBranch     string `json:"default_branch"`
 	Homepage          string `json:"homepage"`
@@ -260,25 +260,25 @@ type Repository struct {
 
 // ObjectAttributes contais all of the GitLab object attributes information
 type ObjectAttributes struct {
-	ID              int        `json:"id"`
+	ID              int64      `json:"id"`
 	Title           string     `json:"title"`
-	AssigneeID      int        `json:"assignee_id"`
-	AuthorID        int        `json:"author_id"`
-	ProjectID       int        `json:"project_id"`
+	AssigneeID      int64      `json:"assignee_id"`
+	AuthorID        int64      `json:"author_id"`
+	ProjectID       int64      `json:"project_id"`
 	CreatedAt       customTime `json:"created_at"`
 	UpdatedAt       customTime `json:"updated_at"`
-	Position        int        `json:"position"`
+	Position        int64      `json:"position"`
 	BranchName      string     `json:"branch_name"`
 	Description     string     `json:"description"`
-	MilestoneID     int        `json:"milestone_id"`
+	MilestoneID     int64      `json:"milestone_id"`
 	State           string     `json:"state"`
-	IID             int        `json:"iid"`
+	IID             int64      `json:"iid"`
 	URL             string     `json:"url"`
 	Action          string     `json:"action"`
 	TargetBranch    string     `json:"target_branch"`
 	SourceBranch    string     `json:"source_branch"`
-	SourceProjectID int        `json:"source_project_id"`
-	TargetProjectID int        `json:"target_project_id"`
+	SourceProjectID int64      `json:"source_project_id"`
+	TargetProjectID int64      `json:"target_project_id"`
 	StCommits       string     `json:"st_commits"`
 	MergeStatus     string     `json:"merge_status"`
 	Content         string     `json:"content"`
@@ -291,13 +291,13 @@ type ObjectAttributes struct {
 	BeforeSHA       string     `json:"before_sha"`
 	Status          string     `json:"status"`
 	Stages          []string   `json:"stages"`
-	Duration        int        `json:"duration"`
+	Duration        int64      `json:"duration"`
 	Note            string     `json:"note"`
 	NotebookType    string     `json:"noteable_type"`
 	At              customTime `json:"attachment"`
 	LineCode        string     `json:"line_code"`
 	CommitID        string     `json:"commit_id"`
-	NoteableID      int        `json:"noteable_id"`
+	NoteableID      int64      `json:"noteable_id"`
 	System          bool       `json:"system"`
 	WorkInProgress  bool       `json:"work_in_progress"`
 	StDiffs         []StDiff   `json:"st_diffs"`
@@ -309,22 +309,22 @@ type ObjectAttributes struct {
 
 // MergeRequest contais all of the GitLab merge request information
 type MergeRequest struct {
-	ID              int        `json:"id"`
+	ID              int64      `json:"id"`
 	TargetBranch    string     `json:"target_branch"`
 	SourceBranch    string     `json:"source_branch"`
 	SourceProjectID string     `json:"source_project_id"`
-	AssigneeID      int        `json:"assignee_id"`
-	AuthorID        int        `json:"author_id"`
+	AssigneeID      int64      `json:"assignee_id"`
+	AuthorID        int64      `json:"author_id"`
 	Title           string     `json:"title"`
 	CreatedAt       customTime `json:"created_at"`
 	UpdatedAt       customTime `json:"updated_at"`
-	MilestoneID     int        `json:"milestone_id"`
+	MilestoneID     int64      `json:"milestone_id"`
 	State           string     `json:"state"`
 	MergeStatus     string     `json:"merge_status"`
-	TargetProjectID int        `json:"target_project_id"`
-	IID             int        `json:"iid"`
+	TargetProjectID int64      `json:"target_project_id"`
+	IID             int64      `json:"iid"`
 	Description     string     `json:"description"`
-	Position        int        `json:"position"`
+	Position        int64      `json:"position"`
 	LockedAt        customTime `json:"locked_at"`
 	Source          Source     `json:"source"`
 	Target          Target     `json:"target"`
@@ -361,7 +361,7 @@ type Source struct {
 	GitSSHURL         string `json:"git_ssh_url"`
 	GitHTTPURL        string `json:"git_http_url"`
 	Namespace         string `json:"namespace"`
-	VisibilityLevel   int    `json:"visibility_level"`
+	VisibilityLevel   int64  `json:"visibility_level"`
 	PathWithNamespace string `json:"path_with_namespace"`
 	DefaultBranch     string `json:"default_branch"`
 	Homepage          string `json:"homepage"`
@@ -379,7 +379,7 @@ type Target struct {
 	GitSSHURL         string `json:"git_ssh_url"`
 	GitHTTPURL        string `json:"git_http_url"`
 	Namespace         string `json:"namespace"`
-	VisibilityLevel   int    `json:"visibility_level"`
+	VisibilityLevel   int64  `json:"visibility_level"`
 	PathWithNamespace string `json:"path_with_namespace"`
 	DefaultBranch     string `json:"default_branch"`
 	Homepage          string `json:"homepage"`
